pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my_project"
version = "0.1.0"
description = "A Twilio and OpenAI GPT integration"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "twilio>=8.12.0",
]

[project.scripts]
my-project = "my_project.main:run_app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 